backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
pandas==2.2.3
numpy==2.1.1
cryptography==43.0.1
python-multipart==0.0.12
anthropic==0.40.0
garage61api==0.1.0
